refactor: change file name from ScrapButton to LikeButton
